refactor: move Form component to base-gallery module
